Rename Blocks renderer
